Refactorings
* Give templates and classes for shopping cart
their own files.
* Rename count to quantity.
* Use totalCount / totalPrice as properties.
