+++ src/app/model/db-shopping-cart.ts
@@ -1,0 +1,14 @@
+import { DbEntry } from "./db-entry";
+import { ShoppingCartData, ShoppingCartProduct } from "./shopping-cart";
+
+/**
+ * Database model for a shopping cart entry.
+ */
+ export interface DbShoppingCart extends ShoppingCartData, DbEntry {
+}
+
+/**
+ * Database model for products within a shopping cart.
+ */
+export interface DbShoppingCartProduct extends ShoppingCartProduct, DbEntry {
+}
