+++ src/app/services/database/shopping-cart-product.service.ts
@@ -1,5 +1,6 @@
 import { Firestore } from '@angular/fire/firestore';
-import { DbShoppingCartProduct } from 'src/app/model/shopping-cart';
+import { DbShoppingCartProduct } from 'src/app/model/db-shopping-cart';
+
 import { AbstractCrudService } from './abstract-crud.service';
 
 export class ShoppingCartProductService extends AbstractCrudService<DbShoppingCartProduct> {
