+++ src/app/services/database/shopping-cart.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Firestore } from '@angular/fire/firestore';
-import { DbShoppingCart } from 'src/app/model/shopping-cart';
+import { DbShoppingCart } from 'src/app/model/db-shopping-cart';
 
 import { AbstractCrudService } from './abstract-crud.service';
